requirements.txt
streamlit
openai
requests
pandas
fastapi
uvicorn
python-dotenv